+++ src/lib/strategies/Unidirectional.ts
@@ -5,8 +5,8 @@ import Mappings, { MappingSnapshot } from '../Mappings'
 import { Folder, ItemLocation, TItem, TItemLocation } from '../Tree'
 import Logger from '../Logger'
 import { CancelledSyncError } from '../../errors/Error'
-import MergeSyncProcess from './Merge'
 import TResource, { IResource, OrderFolderResource } from '../interfaces/Resource'
+import Scanner from '../Scanner'
 
 export default class UnidirectionalSyncProcess extends DefaultStrategy {
   protected direction: TItemLocation
@@ -19,8 +19,47 @@ export default class UnidirectionalSyncProcess extends DefaultStrategy {
     this.direction = direction
   }
 
+  countPlannedActions() {
+    this.actionsPlanned = this.revertPlan.getActions().length
+    this.actionsPlanned += this.revertPlan.getActions(ActionType.CREATE).map(action => action.payload.count()).reduce((a, i) => a + i, 0)
+  }
+
   async getDiffs():Promise<{localDiff:Diff, serverDiff:Diff}> {
-    return MergeSyncProcess.prototype.getDiffs.apply(this) // cheeky!
+    const mappingsSnapshot = this.mappings.getSnapshot()
+
+    const newMappings = []
+    const localScanner = new Scanner(
+      this.serverTreeRoot,
+      this.localTreeRoot,
+      (serverItem, localItem) => {
+        if (localItem.type === serverItem.type && (serverItem.canMergeWith(localItem) || Mappings.mappable(mappingsSnapshot, serverItem, localItem))) {
+          newMappings.push([localItem, serverItem])
+          return true
+        }
+        return false
+      },
+      this.preserveOrder,
+      false
+    )
+    const serverScanner = new Scanner(
+      this.localTreeRoot,
+      this.serverTreeRoot,
+      (localItem, serverItem) => {
+        if (serverItem.type === localItem.type && (serverItem.canMergeWith(localItem) || Mappings.mappable(mappingsSnapshot, serverItem, localItem))) {
+          newMappings.push([localItem, serverItem])
+          return true
+        }
+        return false
+      },
+      this.preserveOrder,
+      false
+    )
+    const [localDiff, serverDiff] = await Promise.all([localScanner.run(), serverScanner.run()])
+    await Promise.all(newMappings.map(([localItem, serverItem]) => {
+      this.addMapping(this.server, localItem, serverItem.id)
+    }))
+
+    return {localDiff, serverDiff}
   }
 
   async loadChildren() :Promise<void> {