+++ src/ui/views/Update.vue
@@ -18,6 +18,30 @@
           </v-btn>
         </v-card-text>
         <v-card-title>
+          {{ t("LabelTelemetry") }}
+        </v-card-title>
+        <v-card-text>
+          <div class="body-1">
+            {{ t("DescriptionTelemetry") }}
+          </div>
+          <v-radio-group v-model="telemetry" class="mt-4">
+            <v-radio :value="true">
+              <template #label>
+                <div class="heading">
+                  {{ t("LabelTelemetryenable") }}
+                </div>
+              </template>
+            </v-radio>
+            <v-radio :value="false">
+              <template #label>
+                <div class="heading">
+                  {{ t("LabelTelemetrydisable") }}
+                </div>
+              </template>
+            </v-radio>
+          </v-radio-group>
+        </v-card-text>
+        <v-card-title>
           {{ t("LabelFunddevelopment") }}
         </v-card-title>
         <v-card-text>
@@ -59,12 +83,14 @@
 <script>
 import {version as VERSION} from '../../../package.json'
 import { routes } from '../NativeRouter'
+import browser from '../../lib/browser-api'
 
 export default {
   name: 'Update',
   components: {},
   data() {
     return {
+      telemetry: false,
       paymentOptions: [
         {
           href: 'https://www.paypal.me/marcelklehr1',
@@ -107,8 +133,15 @@ export default {
       return routes
     }
   },
-  methods: {
-  }
+  watch: {
+    telemetry(enabled) {
+      browser.storage.local.set({'telemetryEnabled': enabled})
+    }
+  },
+  async created() {
+    const {telemetryEnabled} = await browser.storage.local.get({'telemetryEnabled': false})
+    this.telemetry = telemetryEnabled
+  },
 }
 </script>